backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets>=13.0,<15
python-dotenv==1.0.0
google-genai==1.0.0
tokenc==0.1.0
arize-phoenix>=4.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.27.0